Create edgeCapture.sv
Quite often your logic needs to react to a change on some control signal. That can be an external input, something saying that another part of the circuit has done it's job and that it is safe to continue. All sorts of scenarios exist that call for a signal generated by one part of a system to be detected by another part or a different system.

Implement a sticky edge detector (or a edge capture) circuit which captures any neg-edge (from 1-0) transition on a 32-bit input signal. The edge detection and capture must be performed on per bit basis for all the 32-bits of the input. All the flops should be positive edge triggered with asynchronous resets (if any).